% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes_from_literature.R
\name{bh.func}
\alias{bh.func}
\title{Implementing BH procedure.}
\usage{
bh.func(pv, q)
}
\arguments{
\item{pv}{p-values.}

\item{q}{desired FDR level.}
}
\value{
A list with the elements
\item{nr}{number of rejections}
\item{th}{the threshold for rejection of the procedure}
\item{de}{the rejections}
}
\description{
Implementing BH procedure.
}
