% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Hns}
\alias{Hns}
\title{A correct version of the Hns function in package ks.}
\usage{
Hns(x, deriv.order = 0)
}
\description{
A correct version of the Hns function in package ks, all definitions are the same as in the original function.
}
