% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pisLAWS.prl}
\alias{pisLAWS.prl}
\title{A parellel computing version of pisLAWS utilizing foreach function.}
\usage{
pisLAWS.prl(s1, pv2, s2, tau = 0.1, h = 0.3)
}
\arguments{
\item{s1}{the covariates to estimate a non-null probability for.}

\item{pv2}{the p-values used for estimation.}

\item{s2}{the covariates corresponding to the p-values pv2.}

\item{tau}{threshold for null p-values.}

\item{h}{bandwidth matrix H.}
}
\value{
Estimated non-null probability for each covariate in s1.
}
\description{
A parellel computing version of pisLAWS utilizing foreach function.
}
