% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{rhoBH.OR}
\alias{rhoBH.OR}
\title{Implementing the oracle rho-BH procedure.}
\usage{
rhoBH.OR(x, mu1, sigma, ws = NULL, q, pis = NULL)
}
\arguments{
\item{x}{the test statistic, whose null distribution is fixed to be N(0,1).}

\item{mu1}{mean vector for the alternative Gaussian distribution.}

\item{sigma}{sigma vector for the alternative Gaussian distribution.}

\item{ws}{weight vector.}

\item{q}{desired FDR level.}

\item{pis}{non-null probability vector.}
}
\value{
A list with the elements
\item{nr}{number of rejections}
\item{th}{the threshold for rejection of the procedure}
\item{de}{the rejections}
\item{rhow}{q_i, weighted rho-values}
\item{fdps}{FDP estimate at each rejection threshold}
}
\description{
This is a simple implementation where null and alaternative distributions are all assumed to be Gaussian.
Null distribution is fixed as N(0,1).
Alternative distribution is N(mu1_i, sigma_i), where mu1_i and sigma_i refer to each element of the input mu1 and sigma.
Rho-BH with other distributions can be implemented through easy modification of this function.
}
