% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{rhoBH}
\alias{rhoBH}
\title{Implementing the data-driven rho-BH procedure.}
\usage{
rhoBH(x, s_aux1, s_aux2, q, mu0 = 0, sig0 = 1)
}
\arguments{
\item{x}{the test statistic, whose null distribution is Gaussian.}

\item{s_aux1}{a m*dim1 matrix for estimation of non-null probability.}

\item{s_aux2}{a m*dim2 matrix for estimation of alternative distributions.}

\item{q}{desired FDR level.}

\item{mu0}{mean vector for the null Gaussian distribution.}

\item{sig0}{sigma vector for the null Gaussian distribution.}
}
\value{
A list with the elements
\item{nr1}{number of rejections, for the first half of data}
\item{th1}{the threshold for rejection of the procedure, for the first half of data}
\item{nr2}{number of rejections, for the second half of data}
\item{th2}{the threshold for rejection of the procedure, for the second half of data}
\item{de}{the rejections fot the whole procedure}
\item{rdsp1}{the index for the first half of data}
\item{rhow1}{q_i, weighted rho-values, for the first half of data}
\item{fdps1}{FDP estimate at each rejection threshold, for the first half of data}
\item{rdsp2}{the index for the second half of data}
\item{rhow2}{q_i, weighted rho-values, for the second half of data}
\item{fdps2}{FDP estimate at each rejection threshold, for the second half of data}
}
\description{
This is a simple implementation where null distributions are all assumed to be Gaussian.
Rho-BH with other null distributions can be implemented through easy modification.
}
